add segment and statement package for assert
