+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/segment/insert/AssignmentAssert.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.insert;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.segment.insert;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
