+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/segment/insert/InsertNamesAndValuesAssert.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.insert;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.segment.insert;
 
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
