+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/segment/parameter/ParameterMarkerAssert.java
@@ -15,32 +15,37 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.orderby;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.segment.parameter;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.SQLStatementAssertMessage;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.orderby.ExpectedOrderBy;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.order.OrderBySegment;
 import org.apache.shardingsphere.test.sql.SQLCaseType;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
 /**
- * Order by assert.
+ * Parameter marker assert.
  *
  * @author zhangliang
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class OrderByAssert {
+public final class ParameterMarkerAssert {
     
     /**
-     * Assert actual order by segment is correct with expected order by.
-     *
+     * Assert parameter markers count.
+     * 
      * @param assertMessage assert message
-     * @param actual actual order by segment
-     * @param expected expected order by
+     * @param actual actual parameter markers count
+     * @param expected expected parameter markers count
      * @param sqlCaseType SQL case type
      */
-    public static void assertIs(final SQLStatementAssertMessage assertMessage, final OrderBySegment actual, final ExpectedOrderBy expected, final SQLCaseType sqlCaseType) {
-        OrderByItemAssert.assertIs(assertMessage, actual.getOrderByItems(), expected, sqlCaseType, "Order by");
+    public static void assertCount(final SQLStatementAssertMessage assertMessage, final int actual, final int expected, final SQLCaseType sqlCaseType) {
+        if (SQLCaseType.Placeholder == sqlCaseType) {
+            assertThat(assertMessage.getText("Parameter markers count assertion error: "), actual, is(expected));
+        } else {
+            assertThat(assertMessage.getText("Parameter markers count assertion error: "), actual, is(0));
+        }
     }
 }
