+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/segment/table/AlterTableAssert.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.table;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.segment.table;
 
 import com.google.common.base.Joiner;
 import lombok.AccessLevel;
